weighted random generator for items
- ActorType now in actor_list.hpp
- rarity info / random items listed in RandomItemGenerator constructor

generate and equip steel breastplate

player can remove equipped gear

replace Actor::Type with something dynamic

- actor type no longer stored in Type enum, but inferred and
  calclulated using actor's state and components

keep player from using up equipment

filter for relevant items in chooseFromInventory

- also change title of menu based on filter

show player pow and def in UI

- player base pow and def shown, with equip
  bonuses shown separately

forgot to initialize equipType

random generation for items

made rarity a float for more precise definition

removed unused cpp file

+++ src/equippable.cpp
@@ -1,6 +1,6 @@
 #include "equippable.hpp"
 
-Equippable::Equippable(int power, int def) : power(power), def(def) {};
+Equippable::Equippable(float power, float def) : power(power), def(def) {};
 
 void Equippable::setPower(float power)          { this->power = power; }
 void Equippable::setDef(float def)              { this->def = def; }
